requirements.txt
flask==2.0.1
flask_socketio
openai
pymysql
redis
tensorflow
tensorflow_hub
numpy
python-dotenv # .env 변수
cryptography #mysql 비밀번호 암호화
eventlet # socketIO 비동기 처리 가능
gunicorn #wsgi서버